[CBR-341] Don't iterate over the keystore
This also does some minor refactoring: most kernel functionality was defined in
submodules, but `applyBlock`, `rollback`, and `newPending` and co where not.
I've moved them to their own submodule now, along with a new .Read module for
read-only access.

This also cleans up some import lists.
